backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg>=0.29.0
aiosqlite==0.19.0
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client (for AI proxy)
httpx==0.26.0

# Redis (rate limiting, caching)
redis==5.0.1

# Utils
python-dateutil==2.8.2
aiofiles==23.2.1

# Development
python-dotenv==1.0.0
